Refactor data migration for FeedbackResponseCommentSearchDocument class
